Moving hard coded functions in package logic
