+++ npccreater/class.txt
