+++ npccreater/gName.txt
